+++ ui/src/app/views/concept-set-details/component.spec.ts
@@ -32,7 +32,7 @@ import {
   findElements,
   simulateClick as simulateClickReact,
   simulateClickNthElement
-} from 'testing/react-test-helpers';
+} from 'testing/react-testing-utility';
 
 import {
   setupModals,
